airflow/requirements.txt
apache-airflow-providers-apache-spark==4.10.0
apache-airflow-providers-common-sql==1.15.0
trino
apache-airflow-providers-docker==3.10.0
